Move blackboard dialog into seperate folder.
