+++ app/lib/blackboard/blackboard_page.dart
@@ -11,9 +11,9 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
 import 'package:sharezone/blackboard/blackboard_card.dart';
-import 'package:sharezone/blackboard/blackboard_dialog.dart';
 import 'package:sharezone/blackboard/blackboard_view.dart';
 import 'package:sharezone/blackboard/blocs/blackboard_page_bloc.dart';
+import 'package:sharezone/blackboard/create_edit_dialog/blackboard_dialog.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
 import 'package:sharezone/navigation/logic/navigation_bloc.dart';
 import 'package:sharezone/navigation/models/navigation_item.dart';
@@ -23,7 +23,7 @@ import 'package:sharezone_widgets/snackbars.dart';
 import 'package:sharezone_widgets/theme.dart';
 import 'package:sharezone_widgets/widgets.dart';
 
-import 'blackboard_dialog.dart';
+import 'create_edit_dialog/blackboard_dialog.dart';
 import 'details/blackboard_details.dart';
 
 /// Open the blackboard page and returns true, if the user added a blackboard item
