+++ app/lib/blackboard/details/blackboard_details.dart
@@ -11,9 +11,9 @@ import 'package:common_domain_models/common_domain_models.dart';
 import 'package:feature_discovery/feature_discovery.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:sharezone/blackboard/blackboard_dialog.dart';
 import 'package:sharezone/blackboard/blackboard_item.dart';
 import 'package:sharezone/blackboard/blackboard_view.dart';
+import 'package:sharezone/blackboard/create_edit_dialog/blackboard_dialog.dart';
 import 'package:sharezone/blackboard/details/blackboard_item_read_by_users_list/blackboard_item_read_by_users_list_page.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
 import 'package:sharezone/comments/comments_gateway.dart';
